yolo/requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.3.56
numpy>=1.24.3
paddlepaddle>=2.4.2
paddleocr>=2.6.1
paho-mqtt>=2.0.0
flask>=2.0.0
Pillow>=9.0.0
matplotlib>=3.7.1
python-dateutil>=2.8.2
tqdm>=4.65.0
psutil>=5.9.0
flask-cors>=3.0.10
flask-socketio>=5.3.0
python-socketio>=5.7.0
python-engineio>=4.3.0
websocket-client>=1.4.0
torch==2.0.1+cu118 --extra-index-url https://mirrors.aliyun.com/pytorch-wheels/cu118
torchvision==0.15.2+cu118 --extra-index-url https://mirrors.aliyun.com/pytorch-wheels/cu118
scikit-learn>=1.2.0
PyYAML>=6.0
requests>=2.28.0
pandas>=1.5.0
seaborn>=0.12.0
# PaddleOCR相关依赖
scikit-image>=0.19.3
imgaug>=0.4.0
pyclipper>=1.3.0
lmdb>=1.4.1
visualdl>=2.5.3
rapidfuzz>=3.0.0
# 可选依赖
shapely>=2.0.1
scipy>=1.10.1
